+++ addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="script.trakt" name="Trakt" version="3.1.9" provider-name="Trakt.tv, Razzeee">
+<addon id="script.trakt" name="Trakt" version="3.1.10" provider-name="Trakt.tv, Razzeee">
 	<requires>
         <import addon="xbmc.python" version="2.24.0"/>
         <import addon="script.module.simplejson" version="3.3.0"/>
@@ -7,7 +7,7 @@
         <import addon="script.module.dateutil" version="2.4.2"/>
         <import addon="script.module.addon.signals" version="0.0.1"/>
 	</requires>
-	<extension point="xbmc.python.script" library="script.py">
+	<extension point="xbmc.python.script" library="defaultscript.py">
 		<provides>executable</provides>
 	</extension>
 	<extension point="xbmc.service" name="trakt.service" library="default.py" start="login" />
