+++ changelog.txt
@@ -1,5 +1,6 @@
 version 3.1.10
   - fix scrobling problem when using yatse to start playback
+  - internal refactoring
 
 version 3.1.9
   - fallback to only title matching if all else fails
