+++ resources/lib/globals.py
