+++ resources/lib/scrobbler.py
@@ -4,10 +4,10 @@
 import xbmc
 import time
 import logging
-import utilities
-import kodiUtilities
+from resources.lib import utilities
+from resources.lib import kodiUtilities
 import math
-from rating import ratingCheck
+from resources.lib.rating import ratingCheck
 
 logger = logging.getLogger(__name__)
 
@@ -82,7 +82,7 @@ class Scrobbler():
         if not kodiUtilities.getSettingAsBool('scrobble_fallback') and 'id' not in self.curVideo and 'video_ids' not in self.curVideo:
             logger.debug('Aborting scrobble to avoid fallback: %s' % (self.curVideo))
             return
-             
+
         if 'type' in self.curVideo:
             logger.debug("Watching: %s" % self.curVideo['type'])
             if not xbmc.Player().isPlayingVideo():
@@ -186,7 +186,7 @@ class Scrobbler():
                     result['movie']['movieid'] = self.curVideo['id']
                 elif utilities.isEpisode(self.curVideo['type']):
                     result['episode']['episodeid'] = self.curVideo['id']
-                
+
             self.__preFetchUserRatings(result)
 
     def __preFetchUserRatings(self, result):
@@ -286,14 +286,14 @@ class Scrobbler():
             logger.debug("scrobble sending show object: %s" % str(self.traktShowSummary))
             logger.debug("scrobble sending episode object: %s" % str(self.curVideoInfo))
             response = self.traktapi.scrobbleEpisode(self.traktShowSummary, self.curVideoInfo, watchedPercent, status)
-            
+
             if (kodiUtilities.getSettingAsBool('scrobble_secondary_title')):
                 logger.debug('[traktPlayer] Setting is enabled to try secondary show title, if necessary.')
                 # If there is an empty response, the reason might be that the title we have isn't the actual show title,
                 # but rather an alternative title. To handle this case, call the Trakt search function.
                 if response is None:
                     logger.debug("Searching for show title: %s" % self.traktShowSummary['title'])
-                    # This text query API is basically the same as searching on the website. Works with alternative 
+                    # This text query API is basically the same as searching on the website. Works with alternative
                     # titles, unlike the scrobble function.
                     newResp = self.traktapi.getTextQuery(self.traktShowSummary['title'], "show", None)
                     if not newResp:
@@ -307,7 +307,7 @@ class Scrobbler():
                         logger.debug("scrobble sending getTextQuery first show object: %s" % str(showObj))
                         # Now we can attempt the scrobble again, using the primary title this time.
                         response = self.traktapi.scrobbleEpisode(showObj, self.curVideoInfo, watchedPercent, status)
-                    
+
             if response is not None:
                 self.__scrobbleNotification(response)
                 logger.debug("Scrobble response: %s" % str(response))
