+++ resources/lib/utilities.py
