+++ tests/test_utilities.py
@@ -3,7 +3,7 @@
 
 import json
 import pytest
-import utilities
+from resources.lib import utilities
 
 def load_params_from_json(json_path):
     with open(json_path) as f:
